+++ src/airwatch/uploadAppToAirwatch/airwatch/tests/awAssignSmartgroup.test.ts
@@ -11,7 +11,10 @@ afterEach(() => {
 });
 
 it('succeeds with one smart group id', async () => {
-  fetchMock.mockResponseOnce(JSON.stringify({ data: { mockData: 'test' } }));
+  fetchMock.mockResponseOnce(JSON.stringify({ data: { mockData: 'test' } }), {
+    status: 200,
+    headers: { 'content-type': 'application/json' },
+  });
   const mockUrl = appendParamsToUrl(`${awBaseUrl}/mam/apps/internal/1/smartgroups/2`);
   await awAssignSmartgroup(1, [2]);
   expect(fetchMock.mock.calls[0]).toEqual([
@@ -25,7 +28,10 @@ it('succeeds with one smart group id', async () => {
 });
 
 it('succeeds with multiple smart group ids', async () => {
-  fetchMock.mockResponse(JSON.stringify({ data: { mockData: 'test' } }));
+  fetchMock.mockResponse(JSON.stringify({ data: { mockData: 'test' } }), {
+    status: 200,
+    headers: { 'content-type': 'application/json' },
+  });
   const mockUrl1 = appendParamsToUrl(`${awBaseUrl}/mam/apps/internal/1/smartgroups/2`);
   const mockUrl2 = appendParamsToUrl(`${awBaseUrl}/mam/apps/internal/1/smartgroups/3`);