+++ src/airwatch/uploadAppToAirwatch/airwatch/tests/awBeginInstall.test.ts
@@ -1,6 +1,6 @@
 import { FetchMock } from 'jest-fetch-mock';
 import { appendParamsToUrl } from '../../../../services/api';
-import { awBaseUrl, awHeaders } from '../../../../utils';
+import { awBaseUrl, awHeaders, awOrganizationGroupId } from '../../../../utils';
 import { awBeginInstall } from '../awBeginInstall';
 
 const fetchMock: FetchMock = global.fetch;
@@ -15,7 +15,7 @@ const mockBody = JSON.stringify({
   DeviceType: '2',
   BundleId: 'com.lego.corp.bundleid',
   ActualFileVersion: '1.0.0',
-  LocationGroupId: 593,
+  LocationGroupId: awOrganizationGroupId,
   PushMode: 'OnDemand',
   UploadViaLink: true,
   SupportedModels: {
@@ -24,7 +24,10 @@ const mockBody = JSON.stringify({
 });
 
 it('returns the id if the request is successful', async () => {
-  fetchMock.mockResponseOnce(JSON.stringify({ Id: { Value: 10 } }));
+  fetchMock.mockResponseOnce(JSON.stringify({ Id: { Value: 10 } }), {
+    status: 200,
+    headers: { 'content-type': 'application/json' },
+  });
   const mockUrl = appendParamsToUrl(`${awBaseUrl}/mam/apps/internal/begininstall`);
   const data = await awBeginInstall(1, 'com.lego.corp.bundleid', '1.0.0', 'testApp');
   expect(data).toStrictEqual(10);