+++ src/airwatch/uploadAppToAirwatch/airwatch/tests/awUploadBlob.test.ts
@@ -1,6 +1,6 @@
 import { FetchMock } from 'jest-fetch-mock';
 import { appendParamsToUrl } from '../../../../services/api';
-import { awBaseUrl, awHeaders, baseAkamaiFileUrl } from '../../../../utils';
+import { awBaseUrl, awHeaders, awOrganizationGroupId, baseAkamaiFileUrl } from '../../../../utils';
 import { awUploadBlob } from '../awUploadBlob';
 
 const fetchMock: FetchMock = global.fetch;
@@ -14,14 +14,17 @@ const fileName = 'fileName';
 
 const uriParams = {
   filename: fileName,
-  organizationgroupid: 593,
+  organizationgroupid: awOrganizationGroupId,
   filelink: `${baseAkamaiFileUrl}/${fileName}`,
   downloadfilefromlink: true,
   accessvia: 'Direct',
 };
 
 it('returns the blob id if the request is successful', async () => {
-  fetchMock.mockResponseOnce(JSON.stringify({ Value: 1 }));
+  fetchMock.mockResponseOnce(JSON.stringify({ Value: 1 }), {
+    status: 200,
+    headers: { 'content-type': 'application/json' },
+  });
   const mockUrl = appendParamsToUrl(`${awBaseUrl}/mam/blobs/uploadblob`, uriParams);
   const data = await awUploadBlob(fileName);
   expect(data).toStrictEqual(1);